[#31] style: 주문 서비스단 기능별 분리
